fix: improve Snake Game
