+++ README.md
@@ -1,0 +1,5 @@
+# FallingSnake
+
+## Requirements
+
+Git, Maven, OpenJDK 11
