+++ src/main/java/com/game/snake/Snake.java
@@ -1,0 +1,16 @@
+package com.game.snake;
+
+import javax.swing.*;
+
+public class Snake extends JFrame {
+
+    public Snake() {
+        add(new Board());
+        setResizable(false);
+        pack();
+        setTitle("Falling snake");
+        setLocationRelativeTo(null);
+        setDefaultCloseOperation(EXIT_ON_CLOSE);
+    }
+
+}
